+++ @navikt/core/react/src/expansion-card/ExpansionCardHeader.tsx
@@ -1,6 +1,7 @@
 import cl from "clsx";
 import React, { forwardRef, useContext } from "react";
 import { ChevronDownIcon } from "@navikt/aksel-icons";
+import { useI18n } from "../util/i18n/i18n.context";
 import { ExpansionCardContext } from "./context";
 
 export interface ExpansionCardHeaderProps
@@ -13,6 +14,7 @@ const ExpansionCardHeader = forwardRef<
   ExpansionCardHeaderProps
 >(({ children, className, ...rest }, ref) => {
   const panelContext = useContext(ExpansionCardContext);
+  const translate = useI18n("global");
 
   if (panelContext === null) {
     console.error(
@@ -26,18 +28,19 @@ const ExpansionCardHeader = forwardRef<
       ref={ref}
       {...rest}
       className={cl("navds-expansioncard__header", className)}
+      data-open={panelContext.open}
     >
       <div className="navds-expansioncard__header-content">{children}</div>
 
       <button
         className="navds-expansioncard__header-button"
-        onClick={() => panelContext.toggleOpen()}
+        onClick={panelContext.toggleOpen}
         type="button"
         aria-expanded={panelContext.open}
       >
         <ChevronDownIcon
           className="navds-expansioncard__header-chevron"
-          title="Vis mer"
+          title={translate("showMore")}
         />
       </button>
     </div>