+++ @navikt/core/react/src/provider/index.ts
@@ -1,2 +1,6 @@
 "use client";
 export { default as Provider, type ProviderProps } from "./Provider";
+export {
+  AkselTheme as UNSAFE_AkselTheme,
+  useAkselTheme as UNSAFE_useAkselTheme,
+} from "./theme/AkselTheme";