+++ @navikt/core/react/src/util/i18n/i18n.context.ts
@@ -1,7 +1,13 @@
+import { Locale } from "date-fns";
 import { useContext } from "react";
 import { LanguageProviderContext } from "../../provider/i18n/LanguageProvider";
 import { get } from "./get";
-import { Component, ComponentTranslation, Translations } from "./i18n.types";
+import {
+  Component,
+  ComponentTranslation,
+  PartialTranslations,
+  Translations,
+} from "./i18n.types";
 
 /**
  * https://regex101.com/r/LYKWi3/1
@@ -15,41 +21,42 @@ type NestedKeyOf<ObjectType extends object> = {
     : `${Key}`;
 }[keyof ObjectType & (string | number)];
 
+export type TFunction<T extends Component> = (
+  keypath: NestedKeyOf<Translations[T]>,
+  replacements?: Record<string, string | number>,
+) => string;
+
 export function useI18n<T extends Component>(
   componentName: T,
   ...local: (ComponentTranslation<T> | undefined)[]
 ) {
   const languageProviderContext = useContext(LanguageProviderContext);
   const i18n = languageProviderContext.translations;
+  const i18nObjects: (PartialTranslations | undefined)[] = [
+    ...local,
+    ...(Array.isArray(i18n)
+      ? i18n.map((t) => t[componentName])
+      : [i18n[componentName]]),
+  ];
 
   /**
    * https://github.com/Shopify/polaris/blob/2115f9ba2f5bcbf2ad15745233501bff2db81ecf/polaris-react/src/utilities/i18n/I18n.ts#L24
    */
-  const translate = (
-    keypath: NestedKeyOf<Translations[T]>,
-    options?: { replacements: Record<string, string | number> },
-  ) => {
-    const text = get(
-      keypath,
-      ...local,
-      ...(Array.isArray(i18n)
-        ? i18n.map((t) => t[componentName])
-        : [i18n[componentName]]),
-    );
+  const translate: TFunction<T> = (keypath, replacements) => {
+    const text = get(keypath, i18nObjects);
 
-    if (options?.replacements) {
+    if (replacements) {
       return text.replace(REPLACE_REGEX, (match) => {
         const replacement = match.substring(1, match.length - 1);
 
-        if (options.replacements[replacement] === undefined) {
-          const replacementData = JSON.stringify(options.replacements);
-
+        if (replacements[replacement] === undefined) {
+          const replacementData = JSON.stringify(replacements);
           throw new Error(
             `Error translating key '${keypath}'. No replacement syntax ({}) found for key '${replacement}'. The following replacements were passed: '${replacementData}'`,
           );
         }
 
-        return options.replacements[replacement] as string; // can also be a number, but JS doesn't mind...
+        return replacements[replacement] as string; // can also be a number, but JS doesn't mind...
       });
     }
 
@@ -57,4 +64,19 @@ export function useI18n<T extends Component>(
   };
 
   return translate;
+}
+
+export function useDateLocale() {
+  const languageProviderContext = useContext(LanguageProviderContext);
+  const i18n = languageProviderContext.translations;
+  const i18nObjects = Array.isArray(i18n)
+    ? i18n.map((t) => t.global)
+    : [i18n.global];
+
+  for (const obj of i18nObjects) {
+    if (obj?.dateLocale) {
+      return obj.dateLocale as Locale;
+    }
+  }
+  throw new Error("dateLocale not found.");
 }