+++ @navikt/core/react/src/util/i18n/locales/nb.ts
@@ -1,12 +1,17 @@
+import { Locale, nb } from "date-fns/locale";
+
 interface TranslationMap {
-  [component: string]:
-    | Record<string, string>
-    | {
-        [subComponent: string]: Record<string, string>;
-      };
+  [component: string]: Record<string, string | Record<string, string> | Locale>;
 }
 
 export default {
+  global: {
+    dateLocale: nb,
+    showMore: "Vis mer",
+    showLess: "Vis mindre",
+    readOnly: "Skrivebeskyttet",
+  },
+
   FileUpload: {
     dropzone: {
       button: "Velg fil",
@@ -42,6 +47,59 @@ export default {
     Removable: {
       /** Will be appended to the accessible name for the button. */
       labelSuffix: "slett",
+    },
+  },
+  ErrorSummary: {
+    heading: "Du må rette disse feilene før du kan fortsette:",
+  },
+  Loader: {
+    title: "Venter…",
+  },
+  Modal: {
+    close: "Lukk",
+  },
+  Pagination: {
+    previous: "Forrige",
+    next: "Neste",
+  },
+  ProgressBar: {
+    progress: "{current} av {max}",
+    progressUnknown:
+      "Fremdrift kan ikke beregnes, antatt tid er {seconds} sekunder.",
+  },
+  Search: {
+    clear: "Tøm",
+    search: "Søk",
+  },
+  Textarea: {
+    /** Screen readers only */
+    maxLength: "Tekstområde med plass til {maxLength} tegn.",
+    charsTooMany: "{chars} tegn for mye",
+    charsLeft: "{chars} tegn igjen",
+  },
+  Timeline: {
+    dateFormat: "dd.MM.yyyy",
+    dayFormat: "dd.MM",
+    monthFormat: "MMM yy",
+    yearFormat: "yyyy",
+    Row: {
+      noPeriods: "Ingen perioder",
+      period: "{start} til {end}",
+    },
+    Period: {
+      success: "Suksess",
+      warning: "Advarsel",
+      danger: "Fare",
+      info: "Info",
+      neutral: "Nøytral", // TODO: Consider making these global
+      period: "{status} fra {start} til {end}",
+    },
+    Pin: {
+      pin: "Pin: {date}",
+    },
+    Zoom: {
+      zoom: "Zoom tidslinjen {start} til {end}",
+      reset: "Tilbakestill tidsperspektiv",
     },
   },
 } satisfies TranslationMap;