+++ aksel.nav.no/website/components/website-modules/examples/__parts/Dekorator.tsx
@@ -14,7 +14,7 @@ function Env() {
   return (
     <div
       id="decorator-env"
-      data-src={`${MILJO_URL}/env?context=privatperson&simple=true&availableLanguages=[{"locale":"nb","url":"https://www.nav.no/person/kontakt-oss"},{"locale":"en","url":"https://www.nav.no/person/kontakt-oss/en/"}]`}
+      data-src={`${MILJO_URL}/env?context=privatperson&simple=true`}
     />
   );
 }