+++ @navikt/aksel-icons/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @navikt/aksel-icons
 
+## 7.5.3
+
+## 7.5.2
+
+## 7.5.1
+
+## 7.5.0
+
+### Minor Changes
+
+- Icons: 🎉 Added FileJpeg, FilePng and MedicalThermometer. ([#3332](https://github.com/navikt/aksel/pull/3332))
+
+## 7.4.3
+
+## 7.4.2
+
+## 7.4.1
+
 ## 7.4.0
 
 ## 7.3.1